typescript slack, refactor router
