+++ packages/backend/src/routers/apiV1Router.ts
@@ -19,6 +19,7 @@ import { passwordResetLinksRouter } from './passwordResetLinksRouter';
 import { projectRouter } from './projectRouter';
 import { savedChartRouter } from './savedChartRouter';
 import { shareRouter } from './shareRouter';
+import { slackRouter } from './slackRouter';
 import { userRouter } from './userRouter';
 
 export const apiV1Router = express.Router();
@@ -148,3 +149,4 @@ apiV1Router.use('/dashboards/:dashboardUuid', dashboardRouter);
 apiV1Router.use('/password-reset', passwordResetLinksRouter);
 apiV1Router.use('/jobs', jobsRouter);
 apiV1Router.use('/share', shareRouter);
+apiV1Router.use('/slack', slackRouter);
